refactor: remove NewVideoContentHeader component and associated tests; update global styles
